Switched from Jblas to JEigen Lib - (Task #178)
Changed Jblas matrix implementation to JEigen library

---
Task #178: Add implementation of Matrix exponential
