+++ de.dlr.sc.virsat.fdir.core/.classpath
@@ -2,7 +2,8 @@
 <classpath>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
-	<classpathentry exported="true" kind="lib" path="externalLib/jblas-1.2.4.jar"/>
 	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="lib" path="externalLib/Jeigen-onefat.jar"/>
+	<classpathentry kind="lib" path="externalLib/jna-4.1.0.jar"/>
 	<classpathentry kind="output" path="target/classes/"/>
 </classpath>
