+++ de.dlr.sc.virsat.fdir.core/build.properties
@@ -1,7 +1,4 @@
-source.. = src/
 output.. = target/classes/
 bin.includes = META-INF/,\
-               .,\
                about.html,\
-               about_files/,\
-               externalLib/jblas-1.2.4.jar
+               about_files/
